scripts/requirements.txt
# Python dependencies for Smartlead API scripts

# HTTP requests
requests>=2.31.0

# Environment variable management
python-dotenv>=1.0.0

# Optional but useful for lead management
# pandas>=2.0.0
# validators>=0.20.0
